+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Dependency-free, type-safe Dart unit conversion library. Decouple specific unit
 types you need to work with from the rest of your code.
 
-## [Unreleased]
+## [2.2.3]
 ### Added
 - Angle measurement type and units
 
@@ -202,6 +202,7 @@ types you need to work with from the rest of your code.
 - Initial package setup
 
 [Unreleased]: https://bitbucket.org/mongoose13/fling-units/commits/
+[2.2.2]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.2.3/
 [2.2.2]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.2.2/
 [2.2.1]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.2.1/
 [2.2.0]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.2.0/